+++ src/main.js
@@ -3,6 +3,9 @@ import { renderArticleList } from './article-list.js';
 import { setCreateArticleBtnClick } from './article-form-modal.js';
 import { addPostFile } from "./article-form-content";
 import { executeValidation } from "./article-form-validation";
+import { setFormSearchElement } from "./animations/header-animations";
+import { setSortList } from "./animations/header-animations";
+import { setAsideNavList } from "./animations/aside-animations";
 
 const articleListElement = document.getElementById('app');
 const ARTICLE_COUNT = 32;
@@ -12,3 +15,6 @@ renderArticleList(articles, articleListElement);
 setCreateArticleBtnClick();
 addPostFile();
 executeValidation();
+setFormSearchElement();
+setSortList();
+setAsideNavList();