+++ CHANGELOG.md
@@ -6,8 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- add-header-aside-animations [WEB-13](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-13)
 - add-form-validation [WEB12](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-12)
 - add-create-modal [WEB-11](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-11)
+- refactor-html [WEB-10](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-10)
 - add-tor-and-delegation-form [WEB-9](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-9)
 - add-delegation-render-article-list [WEB-8](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-8)
 - add-open-modal-and-style-blog [WEB-7](https://webdot.youtrack.cloud/agiles/171-14/current?issue=WEB-7)